moved testdata into one file and created pointer for nodes
