+++ pkg/assembler/backends/neo4j/artifact.go
@@ -1,0 +1,47 @@
+//
+// Copyright 2023 The GUAC Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package neo4jBackend
+
+import (
+	"strings"
+)
+
+// ArtifactNode is a node that represents an artifact
+type artifactNode struct {
+	algorithm string
+	digest    string
+}
+
+func (an *artifactNode) Type() string {
+	return "Artifact"
+}
+
+func (an *artifactNode) Properties() map[string]interface{} {
+	properties := make(map[string]interface{})
+	properties["algorithm"] = an.algorithm
+	properties["digest"] = strings.ToLower(an.digest)
+	return properties
+}
+
+func (an *artifactNode) PropertyNames() []string {
+	fields := []string{"algorithm", "digest"}
+	return fields
+}
+
+func (an *artifactNode) IdentifiablePropertyNames() []string {
+	// An artifact can be uniquely identified by algorithm and digest
+	return []string{"algorithm", "digest"}
+}
