+++ pkg/assembler/backends/neo4j/builder.go
@@ -15,36 +15,6 @@
 
 package neo4jBackend
 
-import (
-	"github.com/guacsec/guac/pkg/assembler"
-)
-
-func registerAllBuilders(client *neo4jClient) error {
-	err := client.registerBuilder("https://github.com/Attestations/GitHubHostedActions@v1")
-	if err != nil {
-		return err
-	}
-	err = client.registerBuilder("https://tekton.dev/chains/v2")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *neo4jClient) registerBuilder(uri string) error {
-	collectedBuilder := builderNode{
-		uri: uri,
-	}
-	assemblerinput := assembler.AssemblerInput{
-		Nodes: []assembler.GuacNode{collectedBuilder},
-	}
-	err := assembler.StoreGraph(assemblerinput, c.driver)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 // builderNode represents the builder
 type builderNode struct {
 	uri string
