+++ pkg/assembler/backends/testing/ingest_cve.go
@@ -20,11 +20,11 @@ import (
 )
 
 func registerAllCVE(client *demoClient) {
-	client.registerCVE("1970", "CVE-2019-13110")
-	client.registerCVE("2001", "CVE-2014-8139")
-	client.registerCVE("1970", "CVE-2014-8140")
-	client.registerCVE("2023", "CVE-2022-26499")
-	client.registerCVE("1970", "CVE-2014-8140")
+	client.registerCVE("2019", "CVE-2019-13110")
+	client.registerCVE("2014", "CVE-2014-8139")
+	client.registerCVE("2014", "CVE-2014-8140")
+	client.registerCVE("2022", "CVE-2022-26499")
+	client.registerCVE("2014", "CVE-2014-8140")
 }
 
 func (c *demoClient) registerCVE(year, id string) {
